@@ -3,7 +3,7 @@ package nl.rsdt.japp.jotial.availability;
         import android.content.Context;
         import android.content.pm.PackageManager;
         import android.os.Build;
-        import android.support.v4.content.ContextCompat;
+        import androidx.core.content.ContextCompat;
 
 /**
  * @author Dingenis Sieger Sinke
