@@ -1,6 +1,6 @@
 package nl.rsdt.japp.jotial.maps.clustering;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.ArrayList;
 import java.util.Collection;
